refactor fixture: break into smaller chunks
and move it to Browser conftest
